pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemma4-pl"
version = "0.1.0"
description = "Polish language adaptation of Gemma 4 E4B, trained on a Polish text corpus."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "Patryk Orwat", email = "patrykorwat@gmail.com" },
]
keywords = ["gemma", "polish", "language-model", "fine-tuning"]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Runtime deps intentionally loose. On the GPU cluster, use
# `requirements-cluster.txt` which pins everything and resolves correctly
# for aarch64 + CUDA 12.9.
dependencies = [
    "transformers>=4.50",
    "accelerate>=0.34",
    "datasets>=3.0",
    "trl>=0.12",
    "sentencepiece>=0.2",
    "tokenizers>=0.20",
    "huggingface_hub>=0.26",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "omegaconf>=2.3",
    "typer>=0.12",
    "rich>=13.7",
    "wandb>=0.18",
    "tqdm>=4.66",
    "orjson>=3.10",
    "fasttext-wheel>=0.9.2",
    "datasketch>=1.6",
    "regex>=2024.9",
    "unidecode>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
    "ipython>=8.0",
]
eval = [
    "lm-eval>=0.4",
]

[project.scripts]
gemma4-pl-prepare = "gemma4_pl.cli:prepare"
gemma4-pl-sft = "gemma4_pl.cli:sft"
gemma4-pl-eval = "gemma4_pl.cli:evaluate"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
